backend/requirements.txt
Django==6.0.2
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-dotenv==1.0.1
requests==2.32.3
openai==1.65.5
scikit-learn==1.3.2
whitenoise==6.8.2
